refactor: MenuBar 하위 컴포넌트들 위치 변경
